Modularize ScienceDirect parsing components
