+++ paperclip/parsers/sites/sciencedirect/__init__.py
@@ -1,0 +1,3 @@
+from .parser import ScienceDirectParser
+
+__all__ = ["ScienceDirectParser"]
